delivery_app/backend/requirements.txt
Flask
Werkzeug
Flask-Cors
Flask-SQLAlchemy
flask_bcrypt
pymysql
flask_jwt_extended
SQLAlchemy
python-dotenv
gunicorn
axios
Flask-Migrate
python-slugify
Flask-Limiter
redis